Moved SRO application fcn to sro_subfcns
